댓글과 작성자 연결

+++ articles/views.py
@@ -179,12 +179,14 @@ def update(request, pk):
         return redirect('articles:index')
 
 
+@login_required
 def comments_create(request, pk):
     article = Article.objects.get(pk=pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -194,7 +196,9 @@ def comments_create(request, pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 def comments_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect('articles:detail', article_pk)
